requirements.txt
connexion[flask,swagger-ui,uvicorn]
psycopg[binary]
SQLAlchemy
python-jose[cryptography]
requests-oauthlib
